+++ source/main.cpp
@@ -798,7 +798,7 @@ LRESULT CALLBACK WndProc(HWND hWnd, UINT message, WPARAM wParam, LPARAM lParam)
 
 			if (!isBitSet)
 			{
-				int vKey = ConvertWindowsKeycodeToProtonVirtualKey(wParam); 
+				int vKey = ConvertWindowsKeycodeToProtonVirtualKey((int)wParam); 
 				GetMessageManager()->SendGUI(MESSAGE_TYPE_GUI_CHAR_RAW, (float)vKey, 1.0f);  
 				if (vKey != wParam)
 				{
@@ -813,7 +813,7 @@ LRESULT CALLBACK WndProc(HWND hWnd, UINT message, WPARAM wParam, LPARAM lParam)
 				{
 
 				
-				int wmCharKey = VKeyToWMCharKey(wParam);
+				int wmCharKey = VKeyToWMCharKey((int)wParam);
 			
 				if (wmCharKey != 0)
 				{
@@ -848,13 +848,13 @@ LRESULT CALLBACK WndProc(HWND hWnd, UINT message, WPARAM wParam, LPARAM lParam)
 			{
 				//repeat key
 				#ifdef C_DONT_USE_WM_CHAR
-				int vKey = ConvertWindowsKeycodeToProtonVirtualKey(wParam); 
+				int vKey = ConvertWindowsKeycodeToProtonVirtualKey((int)wParam);
 				if (vKey != wParam)
 				{
 					bWasChanged = true;
 				}
 				
-				int wmCharKey = VKeyToWMCharKey(wParam);
+				int wmCharKey = VKeyToWMCharKey((int)wParam);
 				if ( !bWasChanged || (wParam < 37 || wParam > 40 )) //filter out the garbage the arrow keys make
 				{
 					if (wmCharKey != 0)
@@ -876,7 +876,7 @@ LRESULT CALLBACK WndProc(HWND hWnd, UINT message, WPARAM wParam, LPARAM lParam)
 #ifdef _DEBUG
 		//	LogMsg("Got key up %d (%c)", (int)wParam, (char)wParam);
 #endif
-			uint32 key = ConvertWindowsKeycodeToProtonVirtualKey(wParam);
+			uint32 key = ConvertWindowsKeycodeToProtonVirtualKey((int)wParam);
 
 			/*
 			if (key == VIRTUAL_KEY_BACK)
@@ -1078,7 +1078,7 @@ LRESULT CALLBACK WndProc(HWND hWnd, UINT message, WPARAM wParam, LPARAM lParam)
 
 #ifndef RT_DONT_DO_MOVE_TIMER_TRICK
 		if (g_timerID == 0)
-			g_timerID = SetTimer(NULL, 0, 33, (TIMERPROC) TimerProc);
+			g_timerID = (uint32)SetTimer(NULL, 0, 33, (TIMERPROC) TimerProc);
 #endif
 		break;
 
@@ -1910,7 +1910,7 @@ int WINAPI WinMain(HINSTANCE hInstance, HINSTANCE hPrevInstance, TCHAR *lpCmdLin
 				Sleep(1);
 				currSystemTime = GetSystemTimeAccurate();
 			}
-			fpsTimer = currSystemTime + (1000.0 / (double(g_fpsLimit)));
+			fpsTimer = (float)(currSystemTime + (1000.0 / (double(g_fpsLimit))));
 		}
 
 		while (!GetBaseApp()->GetOSMessages()->empty())
